+++ re2c/read.h
@@ -11,12 +11,9 @@
 #ifndef R2READ_H
 #define R2READ_H
 
-#include <assert.h>
-
 #include "scan.h"
 
-
-int read_shiftbuf(scanstate *ss);
+ssize_t read_shiftbuf(scanstate *ss);
 
 #endif
